Move jms tests
